worker/requirements.txt
asyncpraw==7.7.1
asyncpg==0.28.0
python-dotenv==1.0.0
aiohttp==3.9.1